Minor updates Some types and SomeButton/SomeText added

+++ ScreenData/src/main/java/com/pv/screendata/objects/SomeColor.kt
@@ -1,8 +1,8 @@
 package com.pv.screendata.objects
 
 data class SomeColor(
-    val red: Int,
-    val green: Int,
-    val blue: Int,
-    val alpha: Int
+    val red: Float,
+    val green: Float,
+    val blue: Float,
+    val alpha: Float
 )
